Added Continuous Stream Support
Creates an async data connection with the gRPC server on the web client, dynamically updating the table with values. Currently not much to look at, but looks good for a preview. It's also pretty slow sending about 10 data points per second (all exceptions since I didn't test on a RPI).

Note to self: Create a service on the server that generates reports for data points per second and average readings when requested.