Woof
Discovered Span type for efficient contigious memory manipulation. Converted some of the logic for the devices to use it.

Created a singleton for the purpose of continuously gathering device data and storing it in a buffer. The singleton will stop gathering data after a specified amount (should make configurable). When the data is requested, the buffer is cleared.

Added a new table type to my test UI page (really needs cleaned up).

Added a new type of gRPC transaction for the new logic.