Add IO Tests, Fix Site CSS, Create efficient calls in device code
After testing every device using multiple methods, I have come to the conclusion that there is a bottleneck.

All methods are very close with the SingleThreadBinaryChunkTest and BinaryChunkWriterTest coming in on average around 1450 datasets per second. The other tests are about 100 datasets behind. That speed difference isn't that great. Each dataset is 48 bytes so we are talking about around 4.8 KBps difference.

Multithreaded code in C++ is rated at around 2450 datasets per second with each dataset taking around 250 bytes. C++ is much faster, but still not fast enough as we require ~4000 datasets per second. Even when changing the size to around 180 bytes, the sample rate was essentially the same.

The next step is to create identical tests that run on a USB to see if there are improvements.