Discovery!!!
I was under the wrong assumption. I had assumed that my code and .Net Core's IO just wasn't up to the task of performant IO on an SD card.

However, I had a suspicion that this was not the case. I created a new test-case where my most performant methodology of storing data in memory was kept, but I did not then write it to disk. I simply read over it.

Doing this resulted in the exact same speed as my two best test-cases!

I kept the arrays the same size, but left them all zeroed out except for the Accelerometer values. The result was ridiculous. I went from 1450 (the maximum speed I reached on the two previously successful tests and the PureMemoryTest) up to 143,000! The binary writing methodology was so fast that I created a 2 GB file in 5 minutes. Success at last!

The slow code is either in the Gyroscope call, the DateTime call, or the CPU temperature call. I will drilldown on these further and likely call them after an arbitrary amount of iterations. The most important data is the accelerometer data, so not as many datapoints are required of the other devices.

I will likely also need to make a buffer-filling loop up to double the sample rate we require and then check against a stopwatch time. If we haven't elapsed past a second, I will wait the remainder of the time (minus perhaps a tenth of a second) before gathering more data. We don't want to use too much power, generate too much heat, or fill up the SD card with too much data.