Success!
After several iterations of tests, this snapshot serves as the official success. The current state of BinaryChunkWriterTest samples at ~49k SPS. The gyroscope iterates every 200 samples while the accelerometers, time, and cpu temp samples every iteration.

The conversion logic from binary was flawed. It is now fixed and also scales the gyroscope. Visual verification of all the values so far are also a success. I'm all but certain of the gyroscope values. However, we still need to research the accelerometers to be sure that the values we are seeing at rest are to be expected.

I'll need to create a new web page to view data files like these.