Fixed MemoryStram logic errors, Discovered true speed of devices
The true speed of the Accelerometers is not, in fact, bottlenecked by the software. Instead, it is bottlenecked by the hardware. at 1Mhz, the ADC itself is only capable of 8,264 SPS (tested repeatedly with the given MCP3208 library and my own custom library that reduces the allocations dramatically.

At 2Mhz (the maximum of both the MCP3208 and the ADIS16460), the SPS is between 12 and 13 thousand SPS. Here there may be bottlenecking issues. My custom code generally does better than the original MCP code by around 500 SPS.

For byte conversion, I am using a struct as a union rather than BitConverter.

The Gyroscope seems to mess with the SPI line. Whenever a switch occurrs (currently best operated at every 200 iterations, the samples for the accelerometer afterward devicate from their otherwise uniform values. The documentation of the MCP3208 suggests that inconsistencies can occur if the device isn't constantly read from. More research should be done on this event.