Another discovery - MCP3208 Requirement; UI Updates
The ADC requires constant reading from its registers at a consistent speed. Deviating causes the values to begin to attenuate. This was discovered by analyzing the results after performing a multitude of tests.

I believe the results of both the ADC and Gyroscope attenuate further due to the switching of SPI modes when switching between devices. However, I have not tested this thorughly enough to say this is true.

To play it safe, we are getting rid of the PiHat, and have purchased the MCP3208 chip by itself. This chip will be assigned to the second SPI bus on the PI, enabled by the last overlay setting in the config.txt file added tot he repo in this commit.

The reason we are removing the PiHat is becauseit is directly tied to the first SPI bus. Since the second Spi bus does not support SPI Mode 3, it cannot support the Gyroscope. The simplest solutions were to either remove the Gyroscope, or purchase the very cheap DIP version of the MCP. We chose the latter.

Running the ADC at 1.9 Mhz now gives us approximately 15k datapoints per second, giving me some leeway with the code.

I will need to determine the best way to accomodate the device requirements while still maintaining a proper time-stamp. This will require some thought.

The UI has been updated to now  run my performant version of the accelerometer with support for the Gyroscope coming when we upgrade the hardware. The current accelerometer code will also need to update to reflect the correct SPI bus. For the moment, the program runs at the client request given a specified time to run.  After running, details are sent back to the client so the user can decide on how many rows they want to see from the resulting file, and which page, or section, of the file they want the rows from. I may later add the ability to stream the file over so the client can save it. This will likely involve streaming the bytes to a J Interop function, this way the bytes are not maintained in system memory, but are simply passed directly to the user's browser's download directory.

I also made a more performant version of the MCP3208 during my performance test. I was originally suspicious of all the inheritence-heavy methodology the library relies on (and it is fairly heavy). However, even the older RPi 3 processor can handle this fast enough on .Net Core. The slowness factor was actually the clock rate. 8k SPS at 1 Mhz was seemingly the most it could handle. Upping it to 2Mhz nearly doubled the SPS! That was very good to know. My version of the code is still slightly faster, generally aquiring 500 sps more. Furthermore, the Gyroscope is actually very slow. I am unsure wheter this is hardware-specific, or  software-caused. My Gyroscope code will need to be updated to reflect my new MCP3208 code as it is, from what I understand, very performant.