New WPF project for Register Troubleshooting, GyroscopeOnly Updates
Created a new WPF application so I may better and more easilly troubleshoot, configure, and visualize the ADIS16460 registers. Some issues with the initial setup led me to a common issue that I did not run into with my Rover WPF application. I was required to move all GRPC calls to a library. I think I fixed this and can remove it, but since it is working, i'll leave it for now.

Currently, the app is only capable of reading the MSC_CTRL register. There is also a startup write to that register for troubleshooting purposes. It will be removed.

I've learned that the GPIO library suppots interrupts, so I have modified GyroscopeOnly to support the interrupt functionality on the DR pin. It will need to be tested. However, the gyroscope is currently unreliable due to physical configuration issues which has made everything during this update incredibly difficult to understand.

Minor updates to the rest of the project, such as the blazor app's GRPC updates, so that it matches the WPF application's version (newer is better, right?). Also, the ADIS16460 register mapping has been moved to its own file for more organized use.

Added the ability to modify the MSC_CTRL register on the blazor UI.