Footer Cleanup, UART UI Addition
Started working on the UART footer logic. I'm thinking all inject the vibe client inside of the inputmodel for the serial port, but i'm not sure yet.

Cleaned up the footer since I had the chance to do it. Now it is the same across all pages. since it is part of the main layout, it's the same even when navigating to different pages.