Successfully communicated via UART
We had issues communicating via UART due to differences in native voltage between the windows serial device and the Raspberry Pi. Windows was 5V and the pi is 3.3 V, so even though self-tests on both devices worked with no issue, they couldn't communicate with each other.

To completely verify UART functionality, I set up two pis and updated the Vibe2020Tests program with UART functionality.

There are three modes of operation. Send, Receive, and Self-Test (default). To use, simply run the program with either send or recieve after (or leave blank to initiate self-test).

The test between the devices was a success. We can now begin creating the final program that incorporates all devices and UART.