Planning Stages for Telemetry
After about 6 months worth of a break, we've decided to start back up. Right off the bat, we started digging into FFT logic. We watched some tutorials and I've read some of the FFTW documentation so I think I'm starting to understand.

I've created a test that given an input of results, returns a vector of half the size. I'm uncertain  of why since everything i've inferred up until this point should return a vector of equal size. Might just be a misunderstanding.

Either way, a shortcut to verifying the code works even before understanding signal analysis would be to find sample data input and ouput. Then I could pass the input into my algorithm and use Assert to verify the results match the desired output.

To double that effort, I could then verify that FFTW plan can indeed reuse arrays by passing new input into the array and subsequently verify that data.

It's important to note that the purpose of telemetry is solely for backup. That being said our backup must be as sound as our main systems. An unreliable or misplanned backup is summarilly pointless.

On that note, there are severall data points that I think could be useful right away. That is both the input and output of an FFT's statistically relevant data such as min, max, mean, sd, etc.

We can also take the start timestamp of the input data as well as the end timestamp and create a segment ID for generating a proper density chart for easier anlysis.

Furthermore, I've seen some graphs of FFT frequencies and it seems to me that the frequencies could be seen as outliers from curve, which could mean that all we need to do is summarize and report data outside the curve... Perhaps that is the point of FFT analysis. I don't know yet, but I think we're getting close.