November Vibration Tests and Project Updates
Ran two 5 minute tests. One at rest, and one with a single accel axis attached to a motor as well as the gyro.

Ran two 30 second tests. One with two accel axis attached to the motor (forgot to take a snippet of the data per second results).

Had to update the nuget packages due to a RPI 4 update. Before, they kept the same information in the kernal as the RPI 3 pertaining to the CPU. The information was technically incorrect, so they updated. However, the dotnet iot library used that data to determine the board thus breaking some pin number implementations. The update fixed this error.

Due to the update, more degrees conversions were avaible meaning the temperature property names were changed to be more obvious. Had to change some arbitrary lines to reflect this.