I think I finally solved it
Decided to poke around more with FFTW.NET in my unit tests project and then saved the output to a csv to read into a new R file. I then created a graph associating Magnitude on the X axis to the Frequency of the Y axis after speaking with Ron. The results are within expected results! We will do more tests on Friday to verify. We might be close to the conclusion of the Telemetry debacle!