Verified and fixed FFT Results
As it turns out, I was off. I was close, but off. My Frequency bins were incorrect. I had assumed that I needed the Frequencies to come from the results in some shape or fashion. But this is wrong. The Frequencies are really just bins of 1. I have 8000 samples and I had gather them all within 1 second. FS/N = 1 in this case. So, just using the ID of the arrays of data, I was able to cross compare an FFT implementation within R with the original data with my implementation and the results are nearly identical!

It's official. I did it!