Unit Test and R verification for Maximums
Created a unit test to test getting the magnitudes in C# and then verified I could do the same in R. Made graphs of the resulting data to prove that maximums can be used to gather the most important information from a large dataset. Need to play with data summarization a bit more to see if there's more data I could use.