Started Telemetry Modification
Creating processes for Telemetry data analysis and preparation. Trying to maximize performance by allocating the least amount of objects as possible while still keeping processes separate and organized.

Will likely pick at this for a while until I'm satisfied.  The goal is to make the processing as fast as possible so we can get as many processed seconds to ground base as possible. I predict that my current ability with respect to the speed of the FFT library will take approximately 10 seconds for each 1 second of data. If the time exceeds this, I will be unhappy and try to shave as much time off as possible.

I will need to create several testing processes with timers following successful completion of the analysis logic and make it as hands-off as possible so anyone can test the payload.