Create System Approach
Separated the systems from each other in an effort to allow for mock tests and dependency injection. Will need to test to see if I've done it right, or need to refactor. There are a few things I think I'd like to separate out further, but one thing at a time.