Created Background Console Service
Created the initial mockup of our Background Service. This will be used for the final program.

Also dispersed some of the main logic into "System" classes where each system is responsible for a certain segment of logic. This was done for the purpose of creating Mock classes that I've successfully used to run a Mockup of the entire experiment. It can be improved, but it's a good start.