Fixed FFT Oversight on ARM
I made a mistake during my earlier research on FFT libraries that I could use. It turns out that the way FFT.NET chooses which FFTW dll to use is based on architecture type. Namely, it could only discern between x86 and x64. RPI uses ARM.

I was able to get around this by making a fork of FFT.NET and modifying the open source project for my own usecase.

A second part of this was to install two packages on RPI by using the following commands:

sudo apt-get install libfftw3-3
sudo apt-get install libfftw3-dev

The latter downloaded arm-specific lib folders. I hard coded the paths to this lib's two major .so files in FFT.NET's DllImport commands. I didn't know if it would work, but it did! I was able to generate data successfully. A test using the same motor we used on windows would be useful for verifying this code works.

The library for this is no a normal reference and not a nuget package.

On that note, I've modified the background service app so that the systems share the same exact class resources as they were not, before.