Testing with SPS on RPI3
Since we moved back to the RPI3, testing has shown that we recieve far less data points per second.

On average I receive around 4800 SPS. This is still above the 4000 SPS requirement, so no worries there. However, adding SPS as a field in the output of the telemetry program has revealed an issue with the logic.

The way it currently works is we cannot associate a time for each datapoint, and are instead grabbing the time at the end of a chunk of our target SPS. This is due to the I2C clock speed being so low. It is at least 10 time slower than the MCP. I theorize that sometimes our seconds slip by slightly. Unfortunately, I cannot track milliseconds. Our RTC is not that accurate, it would seem. Or I need to create my own protocols for contacting the RTC, which is out of the question at this point in time.

Considering the SPS are pretty consistent, I think i'll continue doing things the way I am. However, in telemetry I may keep track of both the second and the 4800 target SPS. We can probably error correct the timing at a later point by simply coinciding the data points.