Moving onto UART Tests
Now keeping track of SPS during telemetry send logic. This has proven to be a second way of tracking our time since our average SPS is around 4800 SPS on the pi 3. So far, so good.

Moving onto UART tests. Currently, the Pi3 doesn't seem to properly send data through UART. Had this issue with a previous PI3 and assumed it was just the TX pin on that device. The RX pin worked find. However, the same issue is occurring, so more research is needed.