Final Preparations
Modified the code so it runs indefinitely.

Created, enabled and tested the service file for systemd on linux. It will push all console logs to disk.

Modified the conversion file to give statistics. Also gave it some QOL additions to make it more useful to us. Will likely give this tool to WVU and Colorado since they'll be messing with our data.

Ran the service for over an hour and converted the results. Added the zip to TestResults.

Added the final build of the program to a new Builds directory.

Modified the BackgroundService program so that it iterates the directories it creates every time it starts to prevent accidental overwriting.

Modified various project files so that the annoying localization directories are no longer built.