Minor improvements
Added FFTW.Net to calculation library so anyone can run it. When building the library for linux, the custom version of this library will be required. I will create logic for this another time.

Improved comments.

Can now control the lines per file logic in a central variable in the conversion program.